requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
langchain==0.1.0
langchain-google-genai==0.0.11
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.14
sentence-transformers==2.2.2
pydantic==2.5.0
python-dotenv==1.0.0
pypdf==3.17.1
requests==2.31.0
google-generativeai==0.3.0
